32 bit index buffer

+++ apps/model_data.h
@@ -18,7 +18,7 @@ namespace dal {
 
     struct ModelData {
         std::vector<Vertex> m_vertices;
-        std::vector<uint16_t> m_indices;
+        std::vector<uint32_t> m_indices;
     };
